NPC基础设施搭建完成!

+++ nemu/src/cpu/difftest/ref.c
@@ -52,7 +52,7 @@ __EXPORT void difftest_regcpy(void *dut, bool direction)
   }
 }
 
-__EXPORT void difftest_exec(uint64_t n)
+__EXPORT void difftest_exec(uint32_t n)
 {
   cpu_exec(n);
 }
